services/calendar-api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
openai>=1.3.0
httpx>=0.25.0
python-dateutil>=2.8.2
PyJWT>=2.8.0
cryptography

# Error tracking (AMA-225)
sentry-sdk[fastapi]>=2.0.0
